+++ Flowtools/clion-plugin/src/main/kotlin/org/flowcore/clion/FlowcoreFileType.kt
@@ -3,8 +3,8 @@ package org.flowcore.clion
 import com.intellij.openapi.fileTypes.LanguageFileType
 
 object FlowcoreFileType : LanguageFileType(FlowcoreLanguage) {
-    override fun getName() = "Flowcore"
-    override fun getDescription() = "Flowcore source file"
+    override fun getName() = "Lyraform"
+    override fun getDescription() = "Lyraform source file"
     override fun getDefaultExtension() = "flow"
     override fun getIcon() = null
 }