+++ Flowtools/clion-plugin/src/main/kotlin/org/flowcore/clion/FlowcoreLanguage.kt
@@ -2,4 +2,4 @@ package org.flowcore.clion
 
 import com.intellij.lang.Language
 
-object FlowcoreLanguage : Language("Flowcore")
+object FlowcoreLanguage : Language("Lyraform")